DIP conformity for ModuleDtector
