+++ src/main/java/edu/hm/hafner/analysis/ModuleDetector.java
@@ -5,19 +5,13 @@ import java.io.InputStream;
 import java.nio.file.InvalidPathException;
 import java.nio.file.Path;
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.Properties;
-import java.util.Scanner;
-import java.util.jar.Attributes;
-import java.util.jar.Manifest;
-import java.util.regex.Pattern;
 
-import org.apache.commons.io.FilenameUtils;
 import org.apache.commons.lang3.StringUtils;
-import org.xml.sax.SAXException;
 
 import com.google.errorprone.annotations.MustBeClosed;
 
@@ -30,47 +24,11 @@ import edu.hm.hafner.util.PathUtil;
  * @author Christoph Laeubrich (support for OSGi-Bundles)
  */
 public class ModuleDetector {
-    private static final String PLUS = ", ";
     private static final String BACK_SLASH = "\\";
     private static final String SLASH = "/";
-    private static final String ALL_DIRECTORIES = "**/";
 
-    private static final String BUNDLE_VENDOR = "Bundle-Vendor";
-    private static final String BUNDLE_SYMBOLIC_NAME = "Bundle-SymbolicName";
-    private static final String BUNDLE_NAME = "Bundle-Name";
-    private static final String REPLACEMENT_CHAR = "%";
-
-    static final String MAVEN_POM = "pom.xml";
-    static final String ANT_PROJECT = "build.xml";
-    static final String OSGI_BUNDLE = "META-INF/MANIFEST.MF";
-    static final String BUILD_GRADLE = "build.gradle";
-    static final String BUILD_GRADLE_KTS = "build.gradle.kts";
-    static final String SETTINGS_GRADLE = "settings.gradle";
-    static final String SETTINGS_GRADLE_KTS = "settings.gradle.kts";
-
-    private static final String PATTERN = ALL_DIRECTORIES + MAVEN_POM
-            + PLUS + ALL_DIRECTORIES + ANT_PROJECT
-            + PLUS + ALL_DIRECTORIES + OSGI_BUNDLE
-            + PLUS + ALL_DIRECTORIES + BUILD_GRADLE
-            + PLUS + ALL_DIRECTORIES + BUILD_GRADLE_KTS
-            + PLUS + ALL_DIRECTORIES + SETTINGS_GRADLE
-            + PLUS + ALL_DIRECTORIES + SETTINGS_GRADLE_KTS;
-    static final String PLUGIN_PROPERTIES = "plugin.properties";
-    static final String BUNDLE_PROPERTIES = "OSGI-INF/l10n/bundle.properties";
-
-    /**
-     * Detects variations of {@code rootProject.setName($string)}, as well as the Groovy setter shorthand
-     * {@code rootProject.name = $string}.
-     *
-     * @implNote This expression does not guarantee matching left/right quotes or parentheses. For example, it will
-     * match {@code 'Name"}, even though that is not valid Groovy or Kotlin syntax. Similarly, in allowing for Groovy's
-     * parentheses-optional syntax for {@code setName $string}, this pattern makes both open and close parentheses
-     * optional, rather than requiring neither or both, so it will parse an open parenthesis without a matching close,
-     * and vice versa.
-     */
-    private static final Pattern RE_GRADLE_SET_PROJECT_NAME =
-            Pattern.compile("^\\s*rootProject\\.(name\\s*=|setName\\(?)\\s*['\"]([^'\"]*)['\"]\\)?");
-
+    /** A list of all module detectors (e.g. Maven) */
+    private final List<AbstractModuleDetector> moduleDetectors;
     /** The factory to create input streams with. */
     private final FileSystem factory;
     /** Maps file names to module names. */
@@ -88,6 +46,14 @@ public class ModuleDetector {
      */
     public ModuleDetector(final Path workspace, final FileSystem fileSystem) {
         factory = fileSystem;
+
+        moduleDetectors = new ArrayList<>(Arrays.asList(
+                new AntModuleDetector(factory),
+                new GradleModuleDetector(factory),
+                new MavenModuleDetector(factory),
+                new OsgiModuleDetector(factory)
+        ));
+
         fileNameToModuleName = createFilesToModuleMapping(workspace);
         prefixes = new ArrayList<>(fileNameToModuleName.keySet());
         Collections.sort(prefixes);
@@ -104,55 +70,13 @@ public class ModuleDetector {
     private Map<String, String> createFilesToModuleMapping(final Path workspace) {
         Map<String, String> mapping = new HashMap<>();
 
-        String[] projects = find(workspace);
-        collectAntProjects(mapping, projects);
-        collectMavenProjects(mapping, projects);
-        collectGradleProjects(mapping, projects);
-        collectOsgiProjects(mapping, projects);
+        ArrayList<String> projects = find(workspace);
+
+        for (AbstractModuleDetector moduleDetector : moduleDetectors) {
+            moduleDetector.collectProjects(mapping, projects);
+        }
 
         return mapping;
-    }
-
-    private void collectAntProjects(final Map<String, String> mapping, final String... projects) {
-        for (String fileName : projects) {
-            if (fileName.endsWith(ANT_PROJECT)) {
-                addMapping(mapping, fileName, ANT_PROJECT, parseBuildXml(fileName));
-            }
-        }
-    }
-
-    private void collectMavenProjects(final Map<String, String> mapping, final String... projects) {
-        for (String fileName : projects) {
-            if (fileName.endsWith(MAVEN_POM)) {
-                addMapping(mapping, fileName, MAVEN_POM, parsePom(fileName));
-            }
-        }
-    }
-
-    private void collectGradleProjects(final Map<String, String> mapping, final String... projects) {
-        for (String fileName : projects) {
-            if (fileName.endsWith(BUILD_GRADLE) || fileName.endsWith(BUILD_GRADLE_KTS)) {
-                addMapping(mapping, fileName, BUILD_GRADLE, parseGradle(fileName));
-            }
-            else if (fileName.endsWith(SETTINGS_GRADLE) || fileName.endsWith(SETTINGS_GRADLE_KTS)) {
-                addMapping(mapping, fileName, SETTINGS_GRADLE, parseGradleSettings(fileName));
-            }
-        }
-    }
-
-    private void collectOsgiProjects(final Map<String, String> mapping, final String... projects) {
-        for (String fileName : projects) {
-            if (fileName.endsWith(OSGI_BUNDLE)) {
-                addMapping(mapping, fileName, OSGI_BUNDLE, parseManifest(fileName));
-            }
-        }
-    }
-
-    private void addMapping(final Map<String, String> mapping, final String fileName, final String suffix,
-            final String moduleName) {
-        if (StringUtils.isNotBlank(moduleName)) {
-            mapping.put(StringUtils.substringBeforeLast(fileName, suffix), moduleName);
-        }
     }
 
     /**
@@ -183,191 +107,27 @@ public class ModuleDetector {
      *
      * @return the found files (as absolute paths)
      */
-    private String[] find(final Path path) {
-        String[] relativeFileNames = factory.find(path, PATTERN);
-        String[] absoluteFileNames = new String[relativeFileNames.length];
+    private ArrayList<String> find(final Path path) {
+        ArrayList<String> absoluteFileNames = new ArrayList<>();
 
-        for (int file = 0; file < absoluteFileNames.length; file++) {
-            String relativePath = normalizePath(relativeFileNames[file]);
-            if (relativePath.startsWith(SLASH)) {
-                absoluteFileNames[file] = relativePath;
-            }
-            else {
-                absoluteFileNames[file] = new PathUtil().getAbsolutePath(path) + SLASH + relativePath;
+        for (AbstractModuleDetector moduleDetector : moduleDetectors) {
+            String[] relativeFileNames = factory.find(path, moduleDetector.getPattern());
+            for (String relativeFileName : relativeFileNames) {
+                String relativePath = normalizePath(relativeFileName);
+                if (relativePath.startsWith(SLASH)) {
+                    absoluteFileNames.add(relativePath);
+                }
+                else {
+                    absoluteFileNames.add(new PathUtil().getAbsolutePath(path) + SLASH + relativePath);
+                }
             }
         }
+
         return absoluteFileNames;
     }
 
     private String normalizePath(final String fileName) {
         return fileName.replace(BACK_SLASH, SLASH);
-    }
-
-    /**
-     * Returns the project name stored in the build.xml.
-     *
-     * @param buildXml
-     *         Ant build.xml file name
-     *
-     * @return the project name or an empty string if the name could not be resolved
-     */
-    private String parseBuildXml(final String buildXml) {
-        try (InputStream file = factory.open(buildXml)) {
-            SecureDigester digester = new SecureDigester(ModuleDetector.class);
-
-            digester.push(new StringBuilder());
-            String xPath = "project";
-            digester.addCallMethod(xPath, "append", 1);
-            digester.addCallParam(xPath, 0, "name");
-
-            StringBuilder result = digester.parse(file);
-            return result.toString();
-        }
-        catch (IOException | SAXException | InvalidPathException ignored) {
-            // ignore
-        }
-        return StringUtils.EMPTY;
-    }
-
-    /**
-     * Returns the project name estimated from the build.gradle file path.
-     *
-     * @param buildScript
-     *         Gradle build.gradle file path
-     *
-     * @return the project name or an empty string if the name could not be resolved
-     */
-    private String parseGradle(final String buildScript) {
-        String basePath = FilenameUtils.getPathNoEndSeparator(buildScript);
-        String parentDirName = FilenameUtils.getName(basePath);
-        return StringUtils.trimToEmpty(parentDirName);
-    }
-
-    /**
-     * Returns the root project name from the settings.gradle file.
-     *
-     * @param settingsFile
-     *         Gradle settings.gradle file path
-     *
-     * @return the root project override, or an empty string if the name could not be resolved
-     */
-    private String parseGradleSettings(final String settingsFile) {
-        String name = null;
-
-        try (InputStream input = factory.open(settingsFile);
-                Scanner scan = new Scanner(input, "UTF-8")) {
-            while (scan.hasNextLine()) {
-                String line = scan.findInLine(RE_GRADLE_SET_PROJECT_NAME);
-
-                if (line != null) {
-                    name = scan.match().group(2);
-                    break;
-                }
-
-                scan.nextLine();
-            }
-        }
-        catch (IOException | InvalidPathException ignored) {
-            // ignore
-        }
-
-        return StringUtils.defaultIfBlank(name, StringUtils.EMPTY);
-    }
-
-    /**
-     * Returns the project name stored in the POM.
-     *
-     * @param pom
-     *         Maven POM file name
-     *
-     * @return the project name or an empty string if the name could not be resolved
-     */
-    private String parsePom(final String pom) {
-        String name = parsePomAttribute(pom, "name");
-
-        return StringUtils.defaultIfBlank(name, parsePomAttribute(pom, "artifactId"));
-    }
-
-    @SuppressWarnings("OverlyBroadCatchBlock")
-    private String parsePomAttribute(final String pom, final String tagName) {
-        try (InputStream file = factory.open(pom)) {
-            SecureDigester digester = new SecureDigester(ModuleDetector.class);
-            digester.push(new StringBuilder());
-            digester.addCallMethod("project/" + tagName, "append", 0);
-
-            StringBuilder result = digester.parse(file);
-            return result.toString();
-        }
-        catch (IOException | SAXException | InvalidPathException ignored) {
-            // ignore
-        }
-        return StringUtils.EMPTY;
-    }
-
-    /**
-     * Scans a Manifest file for OSGi Bundle Information.
-     *
-     * @param manifestFile
-     *         file name of MANIFEST.MF
-     *
-     * @return the project name or an empty string if the name could not be resolved
-     */
-    private String parseManifest(final String manifestFile) {
-        try (InputStream file = factory.open(manifestFile)) {
-            Manifest manifest = new Manifest(file);
-            Attributes attributes = manifest.getMainAttributes();
-            Properties properties = readProperties(StringUtils.substringBefore(manifestFile, OSGI_BUNDLE));
-            String name = getLocalizedValue(attributes, properties, BUNDLE_NAME);
-            if (StringUtils.isNotBlank(name)) {
-                return name;
-            }
-            return getSymbolicName(attributes, properties);
-        }
-        catch (IOException | InvalidPathException ignored) {
-            // ignore
-        }
-        return StringUtils.EMPTY;
-    }
-
-    private String getLocalizedValue(final Attributes attributes, final Properties properties,
-            final String bundleName) {
-        String value = attributes.getValue(bundleName);
-        if (StringUtils.startsWith(StringUtils.trim(value), REPLACEMENT_CHAR)) {
-            return properties.getProperty(StringUtils.substringAfter(value, REPLACEMENT_CHAR));
-        }
-        return value;
-    }
-
-    private Properties readProperties(final String path) {
-        Properties properties = new Properties();
-        readProperties(path, properties, PLUGIN_PROPERTIES);
-        readProperties(path, properties, BUNDLE_PROPERTIES);
-
-        return properties;
-    }
-
-    @SuppressWarnings("OverlyBroadCatchBlock")
-    private void readProperties(final String path, final Properties properties, final String fileName) {
-        try (InputStream file = factory.open(path + SLASH + fileName)) {
-            properties.load(file);
-        }
-        catch (IOException | InvalidPathException ignored) {
-            // ignore if properties are not present or not readable
-        }
-    }
-
-    private String getSymbolicName(final Attributes attributes, final Properties properties) {
-        String symbolicName = StringUtils.substringBefore(attributes.getValue(BUNDLE_SYMBOLIC_NAME), ";");
-        if (StringUtils.isNotBlank(symbolicName)) {
-            String vendor = getLocalizedValue(attributes, properties, BUNDLE_VENDOR);
-            if (StringUtils.isNotBlank(vendor)) {
-                return symbolicName + " (" + vendor + ")";
-            }
-            else {
-                return symbolicName;
-            }
-        }
-        return StringUtils.EMPTY;
     }
 
     /**
