+++ src/main/java/edu/hm/hafner/analysis/PackageDetectors.java
@@ -22,13 +22,11 @@ class PackageDetectors {
     /** If no package could be assigned this value is used as package name. */
     static final String UNDEFINED_PACKAGE = "-";
 
-    private final List<AbstractPackageDetector> detectors = new ArrayList<>();
+    private final List<AbstractPackageDetector> detectors;
 
     @VisibleForTesting
-    PackageDetectors(final FileSystem fileSystem) {
-        detectors.add(new JavaPackageDetector(fileSystem));
-        detectors.add(new CSharpNamespaceDetector(fileSystem));
-        detectors.add(new KotlinPackageDetector(fileSystem));
+    PackageDetectors(final ArrayList<AbstractPackageDetector> detectors) {
+        this.detectors = detectors;
     }
 
     /**
