+++ src/main/java/edu/hm/hafner/analysis/PackageNameResolver.java
@@ -1,13 +1,15 @@
 package edu.hm.hafner.analysis;
 
 import java.nio.charset.Charset;
+import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.Map;
 import java.util.Set;
 import java.util.stream.Collectors;
 
-import edu.hm.hafner.analysis.PackageDetectors.FileSystem;
 import edu.hm.hafner.util.VisibleForTesting;
 
+import static edu.hm.hafner.analysis.PackageDetectors.*;
 import static java.util.function.Function.*;
 
 /**
@@ -27,7 +29,12 @@ public class PackageNameResolver {
 
     @VisibleForTesting
     PackageNameResolver(final FileSystem fileSystem) {
-        packageDetectors = new PackageDetectors(fileSystem);
+        ArrayList<AbstractPackageDetector> detectors = new ArrayList<>(Arrays.asList(
+                new JavaPackageDetector(fileSystem),
+                new CSharpNamespaceDetector(fileSystem),
+                new KotlinPackageDetector(fileSystem)
+        ));
+        packageDetectors = new PackageDetectors(detectors);
     }
 
     /**
@@ -54,13 +61,12 @@ public class PackageNameResolver {
                 .collect(Collectors.toMap(identity(),
                         fileName -> packageDetectors.detectPackageName(fileName, charset)));
 
-        try (IssueBuilder builder = new IssueBuilder()) {
-            report.stream().forEach(issue -> {
-                if (!issue.hasPackageName()) {
-                    issue.setPackageName(builder.internPackageName(packagesOfFiles.get(issue.getAbsolutePath())));
-                }
-            });
-        }
+        IssueBuilder builder = new IssueBuilder();
+        report.stream().forEach(issue -> {
+            if (!issue.hasPackageName()) {
+                issue.setPackageName(builder.internPackageName(packagesOfFiles.get(issue.getAbsolutePath())));
+            }
+        });
         report.logInfo("-> resolved package names of %d affected files", filesWithoutPackageName.size());
     }
 }
